removed popup screen as it caused problems
