+++ src/contexts/SupabaseAuthContext.tsx
@@ -4,7 +4,6 @@ import { supabase } from '@/integrations/supabase/client';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'sonner';
 import { linkPendingAssessmentToUser } from '@/utils/dnaAssessmentUtils';
-import ExistingAssessmentDialog from '@/components/dna/ExistingAssessmentDialog';
 
 // Define profile interface to include vanity_url
 interface Profile {
@@ -65,11 +64,6 @@ export function AuthProvider({ children }: AuthProviderProps) {
   const [error, setError] = useState<AuthError | null>(null);
   const navigate = useNavigate();
   
-  // State for existing assessment dialog
-  const [showExistingAssessmentDialog, setShowExistingAssessmentDialog] = useState(false);
-  const [existingAssessmentId, setExistingAssessmentId] = useState<string>('');
-  const [pendingAssessmentId, setPendingAssessmentId] = useState<string>('');
-
   useEffect(() => {
     // Get initial session
     supabase.auth.getSession().then(({ data: { session }, error: sessionError }) => {
@@ -90,20 +84,11 @@ export function AuthProvider({ children }: AuthProviderProps) {
                 // Also refresh DNA status
                 checkDNAStatus().catch(console.error);
               } else if (result.hasExistingAssessment && result.existingAssessmentId) {
-                // User already has a DNA assessment, show dialog
+                // User already has a DNA assessment
                 console.log('User already has an existing assessment:', result.existingAssessmentId);
                 
                 // Set the hasCompletedDNA flag
                 setHasCompletedDNA(true);
-                
-                // Store assessment IDs for the dialog
-                setExistingAssessmentId(result.existingAssessmentId);
-                const pendingId = localStorage.getItem('pending_dna_assessment_id');
-                if (pendingId) {
-                  setPendingAssessmentId(pendingId);
-                  // Show the dialog
-                  setShowExistingAssessmentDialog(true);
-                }
               }
             })
             .catch(err => console.error('Error linking assessment on session init:', err));
@@ -126,20 +111,11 @@ export function AuthProvider({ children }: AuthProviderProps) {
               // Also refresh DNA status
               checkDNAStatus().catch(console.error);
             } else if (result.hasExistingAssessment && result.existingAssessmentId) {
-              // User already has a DNA assessment, show dialog
+              // User already has a DNA assessment
               console.log('User already has an existing assessment:', result.existingAssessmentId);
               
               // Set the hasCompletedDNA flag
               setHasCompletedDNA(true);
-              
-              // Store assessment IDs for the dialog
-              setExistingAssessmentId(result.existingAssessmentId);
-              const pendingId = localStorage.getItem('pending_dna_assessment_id');
-              if (pendingId) {
-                setPendingAssessmentId(pendingId);
-                // Show the dialog
-                setShowExistingAssessmentDialog(true);
-              }
             }
           })
           .catch(err => console.error('Error linking assessment on auth change:', err));
@@ -321,12 +297,6 @@ export function AuthProvider({ children }: AuthProviderProps) {
   return (
     <AuthContext.Provider value={value}>
       {children}
-      <ExistingAssessmentDialog 
-        open={showExistingAssessmentDialog}
-        onOpenChange={setShowExistingAssessmentDialog}
-        existingAssessmentId={existingAssessmentId}
-        pendingAssessmentId={pendingAssessmentId}
-      />
     </AuthContext.Provider>
   );
 }
